requirements.txt
# Core ML
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.3.2

# API
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6

# Monitoring
prometheus-client==0.20.0

# Data versioning
dvc==3.48.4
pathspec<0.12.0

# Utilities (Infrastructure/Validation)
pyyaml==6.0.1
requests==2.31.0
